Fix path exists error

+++ setup.py
@@ -58,12 +58,17 @@ except CalledProcessError as err:
 url1 = "https://www.python.org/ftp/python/3.7.4/python-3.7.4-embed-win32.zip"
 url2 = "https://www.python.org/ftp/python/3.7.4/python-3.7.4-embed-amd64.zip"
 local_embed_zip = os.path.join(cache_files_dir, "embed.zip")
-os.unlink(local_embed_zip)
+
+if os.path.exists(local_embed_zip):
+    os.unlink(local_embed_zip)
+
 if not os.path.exists(local_embed_zip):
     print("Downloading a python embed zip...")
     urllib.request.urlretrieve(url2, local_embed_zip)
 
-shutil.rmtree("__utils")
+if os.path.exists("__utils"):
+    shutil.rmtree("__utils")
+
 try:
    check_output(['unzip', '-q', local_embed_zip, '-d', "__utils"], stderr=STDOUT)
 except CalledProcessError as err:
